server/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
faiss-cpu